make GUI useflag optional

+++ app-portage/sisyphus/sisyphus-0.1-r3.ebuild
@@ -18,26 +18,27 @@ EGIT_COMMIT="97ac8f5f4b6f36db3a52cc6a5f72b616127b3ba2"
 LICENSE="GPL-2"
 SLOT="0"
 KEYWORDS="amd64 x86"
-IUSE=""
+IUSE="gui"
 
 DEPEND="dev-lang/python:3.4[sqlite]"
 RDEPEND="${DEPEND}
 	app-portage/gentoolkit[${PYTHON_USEDEP}]
-	dev-python/PyQt5[designer,gui,widgets,${PYTHON_USEDEP}]
 	dev-python/urllib3[${PYTHON_USEDEP}]
-	sys-apps/portage[${PYTHON_USEDEP}]"
+	sys-apps/portage[${PYTHON_USEDEP}]
+	gui? ( dev-python/PyQt5[designer,gui,widgets,${PYTHON_USEDEP}] )"
 
 src_install() {
 	default
 	dosym /usr/bin/${PN}-cli.py /usr/bin/${PN}
 	dodir /var/lib/${PN}/{csv,db}
-	# don't include any graphical interface code...for now
-	rm -rf ${ED}usr/bin/${PN}-gui
-	rm -rf ${ED}usr/bin/${PN}-gui-pkexec
-	rm -rf ${ED}usr/share/${PN}/*py
-	rm -rf ${ED}usr/share/${PN}/icon
-	rm -rf ${ED}usr/share/${PN}/ui
-	rm -rf ${ED}usr/share/applications
-	rm -rf ${ED}usr/share/pixmaps
-	rm -rf ${ED}usr/share/polkit-1
+	if ! use gui; then
+		rm -rf ${ED}usr/bin/${PN}-gui
+		rm -rf ${ED}usr/bin/${PN}-gui-pkexec
+		rm -rf ${ED}usr/share/${PN}/*py
+		rm -rf ${ED}usr/share/${PN}/icon
+		rm -rf ${ED}usr/share/${PN}/ui
+		rm -rf ${ED}usr/share/applications
+		rm -rf ${ED}usr/share/pixmaps
+		rm -rf ${ED}usr/share/polkit-1
+	fi
 }
